+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 ### Features Added
 - Java 21 for build and runtime. [#995](https://github.com/Consensys/web3signer/pull/995)
 - Electra fork support. [#1020](https://github.com/Consensys/web3signer/pull/1020) and [#1023](https://github.com/Consensys/web3signer/pull/1023)
-- Commit boost API - Get Public Keys. [#1031](https://github.com/Consensys/web3signer/pull/1031)
+- Commit Boost API - Get Public Keys. [#1031](https://github.com/Consensys/web3signer/pull/1031)
+- Commit Boost API - Generate Proxy Keys
 
 ### Bugs fixed
 - Override protobuf-java to 3.25.5 which is a transitive dependency from google-cloud-secretmanager. It fixes CVE-2024-7254.